+++ src/main/java/store/OutputView.java
@@ -20,7 +20,7 @@ public class OutputView {
         System.out.printf("\n현재 %s은(는) 1개를 무료로 더 받을 수 있습니다. 추가하시겠습니까? (Y/N)\n", productName);
     }
 
-    public void promotionImpossibleGuide(String productName, int amount) {
+    public void promotionImpossibleGuide(String productName, long amount) {
         System.out.printf("\n현재 %s %d개는 프로모션 할인이 적용되지 않습니다. 그래도 구매하시겠습니까? (Y/N)\n", productName, amount);
     }
